Dice Roller (Still needs some work)
Tried to follow the flowcharts  and UML diagram to the best of my ability. User can choose as many dice to roll at once as long as they are valid die types (4, 6, 8, 10  12, 20).  Percentages are calculated just based on the the number they rolled divided by the total number of sides for that die.  If  multiple die are rolled, the average of each percentage rolled is used (not sure if this is the correct way to do it).  Created a basic UI for testing purposed.